refactor remove unnecessary dep
